Add importable concern
