+++ app/controllers/assignment_invitations_controller.rb
@@ -80,9 +80,8 @@ class AssignmentInvitationsController < ApplicationController
 
   def configurable_submission?
     repo             = current_submission.github_repository
-    import           = repo.import_progress[:status]
     classroom_branch = repo.branch_present? config_branch
-    import == "complete" && classroom_branch && current_submission.not_configured?
+    repo.imported? && classroom_branch && current_submission.not_configured?
   end
 
   def create_submission
